+++ hutool-core/src/main/java/com/xiaoleilu/hutool/bean/BeanDescCache.java
@@ -1,8 +1,6 @@
 package com.xiaoleilu.hutool.bean;
 
-import java.util.Collections;
-import java.util.Map;
-import java.util.WeakHashMap;
+import com.xiaoleilu.hutool.lang.SimpleCache;
 
 /**
  * Bean属性缓存<br>
@@ -13,26 +11,23 @@ import java.util.WeakHashMap;
 public enum BeanDescCache {
 	INSTANCE;
 	
-	private Map<Class<?>, BeanDesc> bdCache = Collections.synchronizedMap(new WeakHashMap<Class<?>, BeanDesc>());
-	private Map<Class<?>, BeanDesc> ignoreCaseBdCache = Collections.synchronizedMap(new WeakHashMap<Class<?>, BeanDesc>());
+	private SimpleCache<Class<?>, BeanDesc> bdCache = new SimpleCache<>();
 	
 	/**
 	 * 获得属性名和{@link BeanDesc}Map映射
 	 * @param beanClass Bean的类
-	 * @param ignoreCase 是否忽略大小写
 	 * @return 属性名和{@link BeanDesc}映射
 	 */
-	public BeanDesc getBeanDesc(Class<?> beanClass, boolean ignoreCase){
-		return (ignoreCase ? ignoreCaseBdCache : bdCache).get(beanClass);
+	public BeanDesc getBeanDesc(Class<?> beanClass){
+		return bdCache.get(beanClass);
 	}
 	
 	/**
 	 * 加入缓存
 	 * @param beanClass Bean的类
 	 * @param BeanDesc 属性名和{@link BeanDesc}映射
-	 * @param ignoreCase 是否忽略大小写
 	 */
-	public void putBeanDesc(Class<?> beanClass, BeanDesc BeanDesc, boolean ignoreCase){
-		(ignoreCase ? ignoreCaseBdCache : bdCache).put(beanClass, BeanDesc);
+	public void putBeanDesc(Class<?> beanClass, BeanDesc BeanDesc){
+		bdCache.put(beanClass, BeanDesc);
 	}
 }